Remove Purchased data models out of ViewModel
